# backend/docs/swagger.yaml
definitions:
...
  dto.CreateProjectInputDTO:
    properties:
      name:
        type: string
      user_id:
        type: string
    type: object
  dto.CreateProtocolInputDTO:
    properties:
      goal_id:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateProjectInputDTO:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    type: object
  dto.UpdateProtocolInputDTO:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  entity.Project:
    properties:
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      userID:
        type: string
    type: object
  entity.Protocol:
    properties:
      goalID:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /project:
    get:
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Project'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List projects by user
      tags:
      - project
    post:
      consumes:
      - application/json
      description: Creates a project for a user
      parameters:
      - description: Project Data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new project
      tags:
      - project
  /project/{id}:
    delete:
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete project
      tags:
      - project
    get:
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Project'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get project by ID
      tags:
      - project
    put:
      consumes:
      - application/json
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Project Data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProjectInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update project
      tags:
      - project
  /protocol:
    get:
      parameters:
      - description: Goal UUID
        in: query
        name: goal_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Protocol'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List protocols by goal
      tags:
      - protocol
    post:
      consumes:
      - application/json
      parameters:
      - description: Protocol Data
        in: body
        name: protocol
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProtocolInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new protocol
      tags:
      - protocol
  /protocol/{id}:
    delete:
      parameters:
      - description: Protocol UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete protocol
      tags:
      - protocol
    get:
      parameters:
      - description: Protocol UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Protocol'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get protocol by ID
      tags:
      - protocol
    put:
      consumes:
      - application/json
      parameters:
      - description: Protocol UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Protocol Data
        in: body
        name: protocol
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProtocolInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update protocol
      tags:
      - protocol
...
swagger: "2.0"